moved stuff around so we can have pages that are not prestyled
