+++ app/(mainSite)/gamedev/page.tsx
@@ -1,16 +1,16 @@
-import { Box, Center, Flex, Stack, styled } from "../../styled-system/jsx";
+import { Box, Center, Flex, Stack, styled } from "styled-system/jsx";
 
 import Image from "next/image";
 import Link from "next/link";
-import Footer from "../../components/Footer/Footer";
+import Footer from "components/Footer/Footer";
 
-import renpyLogo from "../../public/img/RenpyLogo.png";
-import godotLogo from "../../public/img/GodotIcon.svg";
+import renpyLogo from "/public/img/RenpyLogo.png";
+import godotLogo from "/public/img/GodotIcon.svg";
 // import greetingBackground from "../../public/img/Homepage-Background.svg";
 
-import renpyExampleImg from "../../public/img/RenpyScreenshotExample.png";
-import godotExampleImg from "../../public/img/GodotScreenshotExample.png";
-import { css } from "../../styled-system/css";
+import renpyExampleImg from "/public/img/RenpyScreenshotExample.png";
+import godotExampleImg from "/public/img/GodotScreenshotExample.png";
+import { css } from "styled-system/css";
 
 const textSettings = {
     fontSize: { base: "xl", sm: "2xl", md: "3xl" },
