+++ package.json
@@ -24,6 +24,7 @@
         "@react-spring/web": "^9.7.3",
         "@zag-js/dialog": "x",
         "@zag-js/menu": "x",
+        "@zag-js/popover": "^0.46.0",
         "@zag-js/progress": "x",
         "@zag-js/react": "x",
         "framer-motion": "^11.0.8",
